functional luastub with documentation, closes #8
