+++ stubs/lua5.1-luasec/README.md
@@ -1,0 +1,12 @@
+# How to run:
+
+lua5.1 run.lua (url) (port) (ca_file)
+
+# Dependencies:
+
+lua 5.1 installed
+luarocks installed and used to install
+* luasec (`luarocks install luasec` to command line)
+* luasocket (`luarocks install luasocket` to command line)
+
+Tested on Ubuntu 16.04
